deploy/requirements.txt
pulp==2.7.0
scikit-image==0.21.0
torchvision==0.12.0
pillow==9.2.0
zhconv==1.4.3
numpy==1.23.3
levenshtein==0.21.1
dataclasses==0.8
sacrebleu==1.5.1
datasets==2.14.2
omegaconf==2.0.6
six==1.12.0
nltk==3.8.1
bitarray==2.6.0
pyarrow==12.0.1
wandb==0.15.7
setuptools==65.5.0
simuleval==1.1.0
packaging==21.3
tqdm==4.64.1
requests==2.28.1
timm==0.9.10
gurobipy==10.0.3
